+++ app/models/folio/console_ability.rb
@@ -9,9 +9,16 @@ class Folio::ConsoleAbility
     case account.role
     when "manager"
       can :manage, :all
+      cannot :manage, :sidekiq
       cannot :manage, Folio::Account, role: "superuser"
+      cannot :manage, Folio::Account, role: "administrator"
+    when "administrator"
+      can :manage, :all
+      cannot :manage, :sidekiq
     when "superuser"
       can :manage, :all
     end
+
+    Rails.application.config.folio_console_ability_lambda.call(self, account)
   end
 end