+++ db/migrate/20210225135419_add_file_name_to_search_to_folio_files.rb
@@ -6,7 +6,9 @@ class AddFileNameToSearchToFolioFiles < ActiveRecord::Migration[6.0]
     add_index :folio_files, %[(to_tsvector('simple', folio_unaccent(coalesce("folio_files"."file_name_for_search"::text, ''))))], using: :gin, name: "index_folio_files_on_by_file_name_for_search"
 
     unless reverting?
-      remove_index :folio_files, name: :index_folio_files_on_by_file_name
+      if ActiveRecord::Migration.connection.index_exists? :folio_files, name: :by_file_name
+        remove_index :folio_files, name: :index_folio_files_on_by_file_name
+      end
 
       sql = "UPDATE folio_files SET file_name_for_search = REPLACE(REPLACE(file_name, '-', '{d}'), '_', '{u}') WHERE file_name IS NOT NULL"
       ActiveRecord::Base.connection.execute(sql)