+++ lib/folio/engine.rb
@@ -54,6 +54,8 @@ module Folio
     config.folio_users_after_accept_path = :root_path
     config.folio_users_signed_in_root_path = :root_path
 
+    config.folio_console_ability_lambda = -> (ability, account) { }
+
     config.folio_cookie_consent_configuration = {
       enabled: true,
       cookies: {